storage: fix trace use-after-finish
When a command was found worthy of refurbishment but had already been
refurbished, we were incorrectly keeping the command for returning the error.
This had the effect of letting the "real" later incarnation of the command
use a context which had since been finished by the client (which was
erroneously sent the error).

Fixes #7185.